refactor: unify WorkItem completion authority in app
