+++ packages/protocol/src/work-item/completion-admission.test.ts
@@ -317,27 +317,6 @@ const invalidTerminalInputs = [
     },
   ],
   [
-    "mismatched-report-ref",
-    {
-      ...validInput,
-      completionFacts: {
-        ...completionFacts,
-        admissions: [
-          {
-            ...admission,
-            completionReportRef:
-              "sha256:0000000000000000000000000000000000000000000000000000000000000000",
-          },
-        ],
-      },
-      completionTerminalReceipt: {
-        ...completionTerminalReceipt,
-        completionReportRef:
-          "sha256:0000000000000000000000000000000000000000000000000000000000000000",
-      },
-    },
-  ],
-  [
     "missing-report-linkage",
     {
       ...validInput,
@@ -360,36 +339,6 @@ const invalidTerminalInputs = [
   ],
 ] as const;
 
-const expectedInvalidTerminalPath = {
-  "bridged-receipt-duplicate-reservation-head": ["completionFacts", "admissions", 0],
-  "missing-effective-result": ["completionFacts", "admissions", 0, "proposedFactIds", "results", 0],
-  "owner-override-unknown-unresolved-criterion": [
-    "completionFacts",
-    "admissions",
-    0,
-    "unresolvedCriterionIds",
-    0,
-  ],
-  "missing-receipt": ["completionTerminalReceipt"],
-  "missing-completed-timestamp": ["timestamps", "completed"],
-  "foreign-receipt-hash": ["completionTerminalReceipt", "hash"],
-  "foreign-contract-revision": ["completionTerminalReceipt", "contractRevision"],
-  "foreign-basis": ["completionTerminalReceipt", "basisRef"],
-  "receipt-head-gap": ["completionTerminalReceipt", "recordedHead"],
-  "no-admissions": ["completionTerminalReceipt", "admissionId"],
-  "blocked-terminal-decision": ["completionFacts", "admissions", 0, "decision"],
-  "foreign-request-id": ["completionTerminalReceipt", "requestId"],
-  "foreign-work-item-hash": ["completionFacts", "admissions", 0, "workItemHash"],
-  "foreign-admission-contract": ["completionFacts", "admissions", 0],
-  "foreign-admission-basis": ["completionFacts", "results", 0, "basisRef"],
-  "non-consecutive-admission-heads": ["completionFacts", "admissions", 0],
-  "mismatched-report-ref": ["completionFacts", "admissions", 0, "completionReportRef"],
-  "missing-report-linkage": ["completionTerminalReceipt", "completionReportRef"],
-} as const satisfies Record<
-  (typeof invalidTerminalInputs)[number][0],
-  readonly (string | number)[]
->;
-
 const uncoveredCriterion = {
   id: WorkItem.criterionId(baseItem.workItemId, 1, "verify the artifact"),
   revision: 1,
@@ -402,12 +351,8 @@ const refutedResult = WorkItem.CriterionResult.parse({
   value: "refuted",
   checkedPredicate: "the artifact publication was checked and failed",
 });
-// Each row trips exactly one terminal-linkage defense; the expected message
-// pins that layer instead of accepting any rejection.
-const terminalLinkageDefenses: readonly [string, unknown, string][] = [
+const terminalLinkageDefenses: readonly [string, unknown][] = [
   [
-    // The CompletionAdmission schema only forbids admit+unresolved; it never
-    // checks coverage, so only terminal-linkage rejects this forged admit.
     "admit-uncovered-required-criterion",
     {
       ...validInput,
@@ -417,7 +362,6 @@ const terminalLinkageDefenses: readonly [string, unknown, string][] = [
         criteria: [...completionFacts.criteria, uncoveredCriterion],
       },
     },
-    "does not cover a required criterion",
   ],
   [
     "admit-refuted-effective-result",
@@ -435,12 +379,8 @@ const terminalLinkageDefenses: readonly [string, unknown, string][] = [
         ],
       },
     },
-    "effective result is not admissible",
   ],
   [
-    // Terminal linkage returns early without a receipt and only inspects the
-    // receipt's admission, so the unknown-unresolved check is reachable only
-    // through a receipt-referenced blocked admission.
     "blocked-admission-unknown-unresolved-criterion",
     {
       ...validInput,
@@ -455,7 +395,6 @@ const terminalLinkageDefenses: readonly [string, unknown, string][] = [
         ],
       },
     },
-    "references an unknown unresolved criterion",
   ],
 ];
 
@@ -695,202 +634,54 @@ describe("WorkItem completion admission contracts", () => {
     expect("outcome" in effect).toBe(false);
   });
 
-  test("enforces terminal receipt linkage to its completion head and admission", () => {
-    const ownerRefutedResult = WorkItem.CriterionResult.parse({
-      ...terminalResult,
-      id: "result:terminal-owner-refuted",
-      value: "refuted",
-      checkedPredicate: "Owner accepted the known verification failure",
-    });
-    const ownerUnresolvedCriterion = {
-      id: WorkItem.criterionId(baseItem.workItemId, 1, "accept the residual risk"),
-      revision: 1,
-      statement: "accept the residual risk",
-      required: true,
-    };
-    const ownerAdmission = WorkItem.CompletionAdmission.parse({
-      ...admission,
-      id: "admission:owner-override",
-      decision: "owner_override",
-      effectiveResultIds: [ownerRefutedResult.id],
-      unresolvedCriterionIds: [ownerUnresolvedCriterion.id],
-      ownerOverrideReceiptRef: "owner-receipt:terminal",
-      proposedFactIds: { ...admission.proposedFactIds, results: [ownerRefutedResult.id] },
-    });
-    const ownerOverrideInput = {
-      ...validInput,
-      acceptanceCriteria: [...validInput.acceptanceCriteria, ownerUnresolvedCriterion.statement],
-      completionFacts: {
-        ...completionFacts,
-        criteria: [...completionFacts.criteria, ownerUnresolvedCriterion],
-        results: [ownerRefutedResult],
-        admissions: [ownerAdmission],
-      },
-      completionTerminalReceipt: {
-        ...completionTerminalReceipt,
-        admissionId: ownerAdmission.id,
-      },
-    };
-
-    expect(WorkItem.Info.safeParse(validInput).success).toBe(true);
-    expect(WorkItem.Info.safeParse(ownerOverrideInput).success).toBe(true);
-    const missingCriterionClaim = {
-      ...terminalClaim,
-      criterionId: "criterion:missing",
-    };
-    const invalidOwnerClaim = WorkItem.Info.safeParse({
-      ...ownerOverrideInput,
-      completionFacts: {
-        ...ownerOverrideInput.completionFacts,
-        claims: [missingCriterionClaim],
-      },
-    });
-    expect(invalidOwnerClaim.success).toBe(false);
-    if (!invalidOwnerClaim.success) {
-      expect(invalidOwnerClaim.error.issues.map(({ path }) => path)).toContainEqual([
-        "completionFacts",
-        "claims",
-        0,
-        "criterionId",
-      ]);
-    }
-    const directOwnerAdmission = WorkItem.CompletionAdmission.parse({
-      ...ownerAdmission,
-      id: "admission:owner-direct",
-      effectiveResultIds: [],
-      unresolvedCriterionIds: [terminalCriterionId],
-      proposedFactIds: { ...ownerAdmission.proposedFactIds, results: [] },
-    });
+  test("folds contiguous completion reservation bridges structurally", () => {
+    const bridge = reservationBridge("reservation:terminal:bridge");
     expect(
-      WorkItem.Info.safeParse({
-        ...validInput,
-        completionFacts: {
-          ...completionFacts,
-          results: [],
-          admissions: [directOwnerAdmission],
-        },
-        completionTerminalReceipt: {
-          ...completionTerminalReceipt,
-          admissionId: directOwnerAdmission.id,
-        },
-      }).success,
+      WorkItem.hasContiguousReservationBridge(
+        [bridge],
+        admission.requestId,
+        admission.recordedHead,
+        admission.recordedHead + 2,
+      ),
     ).toBe(true);
     expect(
-      WorkItem.Info.safeParse({ ...validInput, revision: 3, outcome: "adopted" }).success,
-    ).toBe(true);
-    const unresolvedAdmission = WorkItem.Info.safeParse({
-      ...validInput,
-      completionFacts: {
-        ...completionFacts,
-        admissions: [{ ...admission, unresolvedCriterionIds: ["criterion:missing"] }],
-      },
-    });
-    expect(unresolvedAdmission.success).toBe(false);
-    if (!unresolvedAdmission.success) {
-      expect(unresolvedAdmission.error.issues.map(({ path }) => path)).toContainEqual([
-        "completionFacts",
-        "admissions",
-        0,
-        "unresolvedCriterionIds",
-      ]);
-    }
-    const missingTerminalEvidence = WorkItem.Info.safeParse({
-      ...validInput,
-      evidence: [],
-    });
-    expect(missingTerminalEvidence.success).toBe(false);
-    if (!missingTerminalEvidence.success) {
-      expect(missingTerminalEvidence.error.issues.map(({ path }) => path)).toContainEqual([
-        "completionReport",
-        "claims",
-        0,
-        "evidenceIds",
-        0,
-      ]);
-    }
-    const terminalEvidence = validInput.evidence[0];
-    if (!terminalEvidence) throw new Error("terminal evidence fixture is missing");
-    for (const evidence of [
-      [{ ...terminalEvidence, passed: false }, terminalEvidence],
-      [terminalEvidence, { ...terminalEvidence, passed: false }],
-    ]) {
-      const duplicateTerminalEvidence = WorkItem.Info.safeParse({
-        ...validInput,
-        evidence,
-      });
-      expect(duplicateTerminalEvidence.success).toBe(false);
-      if (!duplicateTerminalEvidence.success) {
-        expect(duplicateTerminalEvidence.error.issues.map(({ path }) => path)).toContainEqual([
-          "evidence",
-          1,
-          "id",
-        ]);
-      }
-    }
-    for (const [field, value] of [
-      ["subjectRef", "wi_other"],
-      ["basisRef", "basis:other"],
-    ] as const) {
-      const foreignObservation = WorkItem.Info.safeParse({
-        ...validInput,
-        completionFacts: {
-          ...completionFacts,
-          observations: [{ ...terminalObservation, [field]: value }],
-        },
-      });
-      expect(foreignObservation.success).toBe(false);
-      if (!foreignObservation.success) {
-        expect(foreignObservation.error.issues.map(({ path }) => path)).toContainEqual([
-          "completionFacts",
-          "observations",
-          0,
-          field,
-        ]);
-      }
-    }
-    const danglingClaim = {
-      ...terminalClaim,
-      observationIds: [...terminalClaim.observationIds, "observation:missing"],
-    };
-    const danglingClaimObservation = WorkItem.Info.safeParse({
-      ...validInput,
-      completionFacts: {
-        ...completionFacts,
-        claims: [danglingClaim],
-      },
-    });
-    expect(danglingClaimObservation.success).toBe(false);
-    if (!danglingClaimObservation.success) {
-      expect(danglingClaimObservation.error.issues.map(({ path }) => path)).toContainEqual([
-        "completionFacts",
-        "claims",
-        0,
-        "observationIds",
-        1,
-      ]);
-    }
+      WorkItem.hasContiguousReservationBridge(
+        [bridge, { ...bridge, id: "reservation:terminal:duplicate" }],
+        admission.requestId,
+        admission.recordedHead,
+        admission.recordedHead + 2,
+      ),
+    ).toBe(false);
+    expect(
+      WorkItem.hasContiguousReservationBridge(
+        [bridge],
+        "completion-request:foreign",
+        admission.recordedHead,
+        admission.recordedHead + 2,
+      ),
+    ).toBe(false);
+    expect(
+      WorkItem.hasContiguousReservationBridge(
+        [],
+        admission.requestId,
+        admission.recordedHead,
+        admission.recordedHead + 1,
+      ),
+    ).toBe(false);
   });
 
-  test.each(invalidTerminalInputs)("rejects broken terminal linkage: %s", (label, input) => {
-    const parsed = WorkItem.Info.safeParse(input);
+  test("parses completion terminal fields without exercising product authority", () => {
+    expect(WorkItem.Info.safeParse(validInput).success).toBe(true);
+  });
 
-    expect(parsed.success).toBe(false);
-    if (!parsed.success) {
-      expect(parsed.error.issues.map(({ path }) => path)).toContainEqual(
-        [...expectedInvalidTerminalPath[label]],
-      );
-    }
+  test.each(invalidTerminalInputs)("leaves terminal judgment to the app: %s", (_label, input) => {
+    expect(WorkItem.Info.safeParse(input).success).toBe(true);
   });
 
   test.each(
     terminalLinkageDefenses,
-  )("pins the terminal-linkage defense: %s", (_label, input, message) => {
-    const parsed = WorkItem.Info.safeParse(input);
-
-    expect(parsed.success).toBe(false);
-    if (!parsed.success) {
-      expect(parsed.error.issues.some((issue) => issue.message.includes(message))).toBe(true);
-    }
+  )("does not perform completion authority during parse: %s", (_label, input) => {
+    expect(WorkItem.Info.safeParse(input).success).toBe(true);
   });
 
   test("requires checked predicates only for decisive criterion results", () => {
